+++ client/src/scenes/widgets/UserPostsWidget.jsx
@@ -1,0 +1,22 @@
+const UserPostsWidget = () => {
+  // const getUserPosts = useCallback(
+  //   async (signal) => {
+  //     setIsPostsLoading(true);
+  //     const { pagesCount, postsPage } = await fetchUserPosts(
+  //       userId,
+  //       token,
+  //       limit,
+  //       pageNum,
+  //       signal
+  //     );
+  //     setTotalPageCount(pagesCount);
+  //     if (postsPage) dispatch(setPosts({ posts: postsPage }));
+  //     setIsPostsLoading(false);
+  //   },
+  //   [userId, token, pageNum, limit, dispatch, reloadToggle]
+  // );
+
+  return <div></div>;
+};
+
+export default UserPostsWidget;
