+++ server/routes/posts.js
@@ -16,7 +16,7 @@ const router = express.Router();
 const upload = multer();
 
 /* READ */
-router.get("/:id/feed/:limit/:pageNum", verifyToken, getFeedPosts);
+router.get("/:id/feed/:timestamp/:limit/:pageNum", verifyToken, getFeedPosts);
 router.get(
   "/user/:userId/limit/:limit/page/:pageNum",
   verifyToken,
